requirements.txt
# requirements-dev.txt
pytest==9.0.3
pytest-cov==6.0.0
pytest-asyncio==1.3.0
httpx==0.28.1
mysql-connector-python==9.2.0
passlib==1.7.4
bcrypt==4.3.0